Adds support for serializing bigints.
I went with serializing bigints as an object with a type and value, but I'm open to other ideas for how to serialize them so they are JSON compatible.

I don't know why I needed to change the type of `setInCopy`, but the original type definition appears to have been incorrect.

I largely don't understand what most of the properties in `parseProps` are for, so I just tried to copy other thinsg in a way that seemed reasonable.
In particular, `editable` I set to `!forceReadonly` but this was really just a guess on my part.

I made a few changes to improve the type strictness of `JSONValue`.
If these strictness increases are considered undesirable then everything except the addition of `JSONBigint` can be removed without causing problems.